# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sarah Chatbot API
  description: |
    REST API for the Sarah premium chatbot platform.
    Includes endpoints for widget configuration, chat management, agent dashboard,
    VoIP calls, leads, analytics, AI caller, wallboard, and administration.
  version: 1.0.0
  contact:
    name: Redegal
    url: https://redegal.com
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://sarah.redegal.com
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Agent JWT token obtained from POST /api/agents/login
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for widget/external integrations
  schemas:
...
    WidgetConfig:
      type: object
      properties:
        branding:
          type: object
          properties:
            companyName:
              type: string
              example: Redegal
            logoUrl:
              type: string
            poweredByText:
              type: string
            showPoweredBy:
              type: boolean
        colors:
          type: object
          properties:
            primary:
              type: string
              example: "#007fff"
            gradientFrom:
              type: string
            gradientTo:
              type: string
        features:
          type: object
          properties:
            enableVoip:
              type: boolean
            enableFileUpload:
              type: boolean
            enableCsat:
              type: boolean
            enableLanguageSelector:
              type: boolean
        layout:
          type: object
          properties:
            position:
              type: string
              enum: [bottom-right, bottom-left]
            width:
              type: integer
            maxHeight:
              type: integer
        i18n:
          type: object
          properties:
            defaultLanguage:
              type: string
            availableLanguages:
              type: array
              items:
                type: string
        businessHours:
          type: object
          properties:
            timezone:
              type: string
            schedule:
              type: array
              items:
                type: object
...
    Lead:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        status:
          type: string
          enum: [new, contacted, qualified, converted, lost]
        score:
          type: integer
        businessLine:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [admin, agent, supervisor]
        email:
          type: string
        languages:
          type: array
          items:
            type: string
        businessLines:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [online, busy, away, offline]
...
    Callback:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
        name:
          type: string
        date:
          type: string
          format: date
        timeSlot:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, missed, cancelled]
...
    FileUpload:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
        name:
          type: string
        mimeType:
          type: string
        size:
          type: integer
    AnalyticsStats:
      type: object
      properties:
        totalConversations:
          type: integer
        totalMessages:
          type: integer
        avgCsat:
          type: number
        leadCount:
          type: integer
    WallboardData:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        global:
          type: object
          properties:
            activeCalls:
              type: integer
            activeChats:
              type: integer
            inQueue:
              type: integer
            agentsOnline:
              type: integer
            slaPercent:
              type: integer
        queues:
          type: array
          items:
            type: object
        agents:
          type: array
          items:
            type: object
        alerts:
          type: array
          items:
            type: object
    Error:
      type: object
      properties:
        error:
          type: string
paths:
...
  /api/config/widget:
    get:
      tags: [Config]
      summary: Get widget theme configuration
      description: Public endpoint called by the widget loader to fetch theme, branding, and feature configuration.
      parameters:
        - name: tenant
          in: query
          schema:
            type: string
          description: Tenant slug (defaults to "redegal")
      responses:
        '200':
          description: Widget configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WidgetConfig'
...
  /api/leads:
    get:
      tags: [Leads]
      summary: List leads
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, contacted, qualified, converted, lost]
        - name: businessLine
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Array of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lead'
...
  /api/agents:
    get:
      tags: [Agents]
      summary: List online agents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Array of agents (without password hashes)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
...
  /api/callbacks:
    get:
      tags: [Callbacks]
      summary: List callbacks with filters
      security:
        - BearerAuth: []
      parameters:
        - name: businessLine
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Array of callbacks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Callback'
...
  /api/analytics:
    get:
      tags: [Analytics]
      summary: Get analytics statistics
      security:
        - BearerAuth: []
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Analytics stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsStats'

  # ═══════════════════════════════════════════════════════
  # UPLOAD
  # ═══════════════════════════════════════════════════════
  /api/upload:
    post:
      tags: [Upload]
      summary: Upload a file
      security:
        - ApiKey: []
      parameters:
        - name: conversationId
          in: query
          schema:
            type: string
            format: uuid
        - name: sender
          in: query
          schema:
            type: string
            default: visitor
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUpload'
        '400':
          description: Invalid file type or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: File too large (max 10 MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ═══════════════════════════════════════════════════════
  # SETTINGS
  # ═══════════════════════════════════════════════════════
  /api/settings/setup:
    get:
      tags: [Settings]
      summary: Check if initial setup is needed
      responses:
        '200':
          description: Setup status
          content:
            application/json:
              schema:
                type: object
                properties:
                  setupRequired:
                    type: boolean
    post:
      tags: [Settings]
      summary: Complete initial setup
      description: First-time setup (no auth required). If already completed, requires admin auth.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                smtp:
                  type: object
                  properties:
                    host:
                      type: string
                    port:
                      type: integer
                    user:
                      type: string
                    password:
                      type: string
                    from:
                      type: string
                sip:
                  type: object
                  properties:
                    domain:
                      type: string
                    port:
                      type: integer
                    extension:
                      type: string
                    password:
                      type: string
                ai:
                  type: object
                  properties:
                    provider:
                      type: string
                hours:
                  type: object
                  properties:
                    timezone:
                      type: string
                    start:
                      type: string
                    end:
                      type: string
                brand:
                  type: object
                  properties:
                    primaryColor:
                      type: string
                    companyName:
                      type: string
      responses:
        '200':
          description: Setup completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
  /api/settings:
    get:
      tags: [Settings]
      summary: Get all settings (admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: All settings
          content:
            application/json:
              schema:
                type: object
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Settings]
      summary: Update settings (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings:
                  type: object
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  updated:
                    type: array
                    items:
                      type: string
...
  /api/wallboard:
    get:
      tags: [Wallboard]
      summary: Get wallboard data
      description: Returns complete real-time contact center metrics including global stats, queue status, agent states, alerts, and hourly volume.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Wallboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WallboardData'

  # ═══════════════════════════════════════════════════════
  # TRANSCRIPTS
  # ═══════════════════════════════════════════════════════
  /api/conversations/{id}/transcript:
    get:
      tags: [Transcripts]
      summary: Export conversation transcript
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          schema:
            type: string
            enum: [json, html, text]
            default: json
      responses:
        '200':
          description: Transcript in requested format
          content:
            application/json:
              schema:
                type: object
            text/html:
              schema:
                type: string
            text/plain:
              schema:
                type: string

  # ═══════════════════════════════════════════════════════
  # DOCS
  # ═══════════════════════════════════════════════════════
...