# server/index.js
require('dotenv').config({ path: require('path').resolve(__dirname, '..', '.env') });

const http = require('http');
const express = require('express');
const path = require('path');
const { WebSocketServer } = require('ws');
const { logger, requestLoggerMiddleware } = require('./utils/logger');
const { db } = require('./utils/db');
const { redis } = require('./utils/redis');
const { loadLanguages } = require('./utils/i18n');
const { loadKnowledgeFiles, seedKnowledgeToDB } = require('./services/knowledge-base');
const compression = require('compression');
const { metricsMiddleware } = require('./services/metrics');
const { corsMiddleware } = require('./middleware/cors');
const { securityHeaders } = require('./middleware/security-headers');
const { csrfProtection } = require('./middleware/csrf');
const { requestId } = require('./middleware/request-id');
const { resolveTenant } = require('./middleware/auth');
const { errorHandler, notFoundHandler } = require('./middleware/error-handler');

const app = express();
const server = http.createServer(app);

// ─── Middleware ───
app.use(compression());
app.use(requestId);
app.use(requestLoggerMiddleware);
app.use(metricsMiddleware);
app.use(securityHeaders);
app.use(corsMiddleware);
app.use(express.json({ limit: '100kb' }));
app.use(csrfProtection);
app.use(resolveTenant);

// ─── Static files ───
app.use('/widget', express.static(path.join(__dirname, 'public', 'widget')));
app.use('/dashboard', express.static(path.join(__dirname, 'public', 'dashboard')));
app.use('/css', express.static(path.join(__dirname, 'public', 'css')));
app.use('/js', express.static(path.join(__dirname, 'public', 'js')));
app.use('/uploads', express.static(path.join(__dirname, 'uploads')));

// ─── robots.txt (disallow all — staging only) ───
app.get('/robots.txt', (_req, res) => {
  res.type('text/plain').send('User-agent: *\nDisallow: /\n');
});

// ─── REST Routes ───
app.use(require('./routes/metrics'));
app.use(require('./routes/health'));
app.use('/api/config', require('./routes/config'));
app.use('/api/chat', require('./routes/chat'));
app.use('/api/leads', require('./routes/leads'));
app.use('/api/agents', require('./routes/agents'));
app.use('/api/calls', require('./routes/call'));
app.use('/api/callbacks', require('./routes/callbacks'));
app.use('/api/analytics', require('./routes/analytics'));
app.use('/api/upload', require('./routes/upload'));
app.use('/api/settings', require('./routes/settings'));
app.use('/api/training', require('./routes/training'));
app.use('/api/ai-caller', require('./routes/ai-caller'));
app.use('/api/wallboard', require('./routes/wallboard'));
app.use('/api/docs', require('./routes/docs'));

// ─── WebSocket upgrade handling ───
const wssChat = new WebSocketServer({ noServer: true, maxPayload: 16 * 1024 });
const wssAgent = new WebSocketServer({ noServer: true, maxPayload: 64 * 1024 });
const wssSip = new WebSocketServer({ noServer: true, maxPayload: 16 * 1024 });

server.on('upgrade', (req, socket, head) => {
  // Validate WebSocket origin in production
  const allowedOrigins = (process.env.ALLOWED_ORIGINS || '').split(',').map(o => o.trim()).filter(Boolean);
  const origin = req.headers.origin;
  if (process.env.NODE_ENV === 'production' && allowedOrigins.length && origin && !allowedOrigins.includes(origin)) {
    logger.warn(`WS: Rejected upgrade from unauthorized origin: ${origin}`);
    socket.write('HTTP/1.1 403 Forbidden\r\n\r\n');
    socket.destroy();
    return;
  }

  const { pathname } = new URL(req.url, `http://${req.headers.host}`);
  if (pathname === '/ws/chat') {
    wssChat.handleUpgrade(req, socket, head, (ws) => wssChat.emit('connection', ws, req));
  } else if (pathname === '/ws/agent') {
    wssAgent.handleUpgrade(req, socket, head, (ws) => wssAgent.emit('connection', ws, req));
  } else if (pathname === '/ws/sip') {
    wssSip.handleUpgrade(req, socket, head, (ws) => wssSip.emit('connection', ws, req));
  } else {
    socket.destroy();
  }
});

// ─── Attach WS handlers ───
const { initChatHandler } = require('./ws/chat-handler');
initChatHandler(wssChat);

const { initAgentHandler } = require('./ws/agent-handler');
initAgentHandler(wssAgent);

const { initSipSignaling } = require('./ws/sip-signaling');
initSipSignaling(wssSip);

// ─── SIP RDGPhone (Vozelia Cloud PBX) ───
const { initSipRDGPhone } = require('./services/sip-rdgphone');
initSipRDGPhone();

// ─── Email notifications ───
const { initEmail } = require('./services/email');
initEmail();

// ─── KB Auto-Updater (scrape redegal.com every 24h) ───
const { initScraper } = require('./services/kb-scraper');
initScraper();

// ─── Call Recording Cleanup (30-day retention) ───
const { initRecordingCleanup } = require('./services/call-recording');
initRecordingCleanup();

// ─── Transcript export ───
const { asyncRoute } = require('./middleware/error-handler');
const { requireAgent } = require('./middleware/auth');
const { generateTranscript } = require('./services/transcript-export');

app.get('/api/conversations/:id/transcript', requireAgent, asyncRoute(async (req, res) => {
  const format = req.query.format || 'json';
  const transcript = await generateTranscript(req.params.id);
  if (format === 'html') {
    res.setHeader('Content-Type', 'text/html; charset=utf-8');
    res.setHeader('Content-Disposition', `attachment; filename="transcript-${req.params.id.slice(0, 8)}.html"`);
    return res.send(transcript.html);
  }
  if (format === 'text') {
    res.setHeader('Content-Type', 'text/plain; charset=utf-8');
    res.setHeader('Content-Disposition', `attachment; filename="transcript-${req.params.id.slice(0, 8)}.txt"`);
    return res.send(transcript.text);
  }
  res.json(transcript);
}));

// ─── Setup wizard (redirect if already completed) ───
app.get('/setup', async (_req, res) => {
  const settings = require('./services/settings');
  const done = await settings.isSetupComplete();
  if (done) return res.redirect('/dashboard');
  res.sendFile(path.join(__dirname, 'public', 'setup.html'));
});

// ─── Premium test page (CSP nonce injected) ───
app.get('/widget/test.html', (_req, res) => {
  const nonce = res.locals.cspNonce;
  res.send(`<!DOCTYPE html>
<html lang="es">
<head>
  <meta charset="UTF-8">
  <meta name="viewport" content="width=device-width, initial-scale=1.0">
  <title>Sarah - Widget Demo</title>
  <style>
    * { margin: 0; padding: 0; box-sizing: border-box; }
    body { font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif; background: linear-gradient(135deg, #667eea 0%, #764ba2 100%); min-height: 100vh; }
    .demo-page { max-width: 1200px; margin: 0 auto; padding: 60px 24px; color: white; }
    h1 { font-size: 48px; font-weight: 800; margin-bottom: 16px; }
    p { font-size: 18px; opacity: 0.9; max-width: 600px; line-height: 1.6; }
    .cards { display: grid; grid-template-columns: repeat(auto-fit, minmax(250px, 1fr)); gap: 24px; margin-top: 48px; }
    .card { background: rgba(255,255,255,0.1); backdrop-filter: blur(12px); border: 1px solid rgba(255,255,255,0.2); border-radius: 16px; padding: 24px; }
    .card h3 { font-size: 18px; margin-bottom: 8px; }
    .card p { font-size: 14px; opacity: 0.8; }
    .config-panel { margin-top: 48px; background: rgba(255,255,255,0.1); backdrop-filter: blur(12px); border: 1px solid rgba(255,255,255,0.2); border-radius: 16px; padding: 24px; }
    .config-panel h3 { margin-bottom: 16px; }
    .config-row { display: flex; gap: 12px; flex-wrap: wrap; align-items: center; margin-bottom: 12px; }
    .config-row label { min-width: 120px; font-size: 14px; }
    .config-row input, .config-row select { padding: 6px 12px; border-radius: 8px; border: 1px solid rgba(255,255,255,0.3); background: rgba(255,255,255,0.1); color: white; font-size: 14px; }
    .config-row input[type="color"] { padding: 0; width: 40px; height: 32px; cursor: pointer; }
    .btn { padding: 8px 20px; background: white; color: #764ba2; font-weight: 600; border: none; border-radius: 8px; cursor: pointer; font-size: 14px; }
    .btn:hover { opacity: 0.9; }
  </style>
</head>
<body>
  <div class="demo-page">
    <h1>Sarah</h1>
    <p>Widget premium con chatbot IA, SarahPhone VoIP, 4 idiomas y configuracion completa. Prueba el widget en la esquina inferior derecha.</p>

    <div class="cards">
      <div class="card">
        <h3>4 Idiomas</h3>
        <p>Deteccion automatica + selector manual. ES, EN, PT, GL</p>
      </div>
      <div class="card">
        <h3>4 Lineas de Negocio</h3>
        <p>Boostic (SEO), Binnacle (BI), Marketing, Tech. Routing inteligente por intencion.</p>
      </div>
      <div class="card">
        <h3>IA Multi-Provider</h3>
        <p>Claude (primario) + Gemini (fallback gratuito) + OpenAI. Cambio automatico si falla.</p>
      </div>
      <div class="card">
        <h3>VoIP SarahPhone</h3>
        <p>Llamadas via SIP + Vozelia Cloud PBX. Solo en horario laboral.</p>
      </div>
    </div>

    <div class="config-panel">
      <h3>Personalizar Widget en Tiempo Real</h3>
      <div class="config-row">
        <label>Color primario:</label>
        <input type="color" id="cfg-color" value="#007fff" />
      </div>
      <div class="config-row">
        <label>Posicion:</label>
        <select id="cfg-position">
          <option value="bottom-right">Abajo derecha</option>
          <option value="bottom-left">Abajo izquierda</option>
        </select>
      </div>
      <div class="config-row">
        <label>Idioma:</label>
        <select id="cfg-language">
          <option value="auto">Auto-detectar</option>
          <option value="es">Espanol</option>
          <option value="en">English</option>
          <option value="pt">Portugues</option>
          <option value="gl">Galego</option>
        </select>
      </div>
    </div>
  </div>

  <script nonce="${nonce}">
    window.Sarah = {
      baseUrl: window.location.origin + '/widget',
      apiUrl: 'ws://' + window.location.host + '/ws/chat',
      configUrl: window.location.origin + '/api/config/widget',
      language: 'auto',
      primaryColor: '#007fff'
    };
    window.RdgBot = window.Sarah;
    function updateConfig(key, val) {
      if (window.__redegalWidget) window.__redegalWidget.updateConfig({ [key]: val });
    }
    document.getElementById('cfg-color').addEventListener('change', function() { updateConfig('primaryColor', this.value); });
    document.getElementById('cfg-position').addEventListener('change', function() { updateConfig('position', this.value); });
    document.getElementById('cfg-language').addEventListener('change', function() { updateConfig('language', this.value); });
  </script>
  <script nonce="${nonce}" src="/widget/loader.js" async></script>
</body>
</html>`);
});

// ─── Dashboard SPA catch-all (for client-side routing, e.g. /dashboard/wallboard) ───
app.get('/dashboard/*', (_req, res) => {
  res.sendFile(path.join(__dirname, 'public', 'dashboard', 'index.html'));
});

// ─── Corporate landing page (root) ───
app.get('/', (_req, res) => {
  res.sendFile(path.join(__dirname, 'public', 'index.html'));
});

// ─── Error handling (must be AFTER all routes) ───
app.use(notFoundHandler);
app.use(errorHandler);

// ─── Startup ───
const PORT = process.env.PORT || 3000;

async function start() {
  // DB and Redis are required — fail fast if unavailable
  try {
    await db.connect();
    logger.info('PostgreSQL connected');
  } catch (e) {
    logger.error('PostgreSQL connection failed — aborting startup:', e.message);
    process.exit(1);
  }

  try {
    await redis.connect();
    logger.info('Redis connected');
  } catch (e) {
    logger.error('Redis connection failed — aborting startup:', e.message);
    process.exit(1);
  }

  loadLanguages();
  loadKnowledgeFiles();

  seedKnowledgeToDB().catch((e) => logger.warn('Knowledge seed error:', e.message));

  const HOST = process.env.BIND_HOST || '127.0.0.1';
  server.listen(PORT, HOST, () => {
    logger.info(`Redegal Web + Chatbot running on http://${HOST}:${PORT}`);
    logger.info(`Dashboard: http://${HOST}:${PORT}/dashboard`);
    logger.info(`Health: http://${HOST}:${PORT}/health`);
    logger.info(`Only accessible via localhost / SSH tunnel`);
  });
}

start();

module.exports = { app, server, wssChat, wssAgent, wssSip };